requirements.txt
requests==2.27.1
python-dotenv==1.0.0
python-dateutil==2.8.2
postmarker==1.0
pytz==2023.3
